ingestion/requirements.txt
ebooklib==0.18
ollama>=0.3.0
pypdf>=4.3.0
google-genai>=1.0.0
supabase>=2.5.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
tqdm>=4.66.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
